fix the logic in the target language index function
We added another language but did not increase the range of the
potential second index. Now the function can return an index that is up
to three off the initial index, instead of just up to two.

This change makes that logic dynamic and will work for any number of
languages.